Fixed initial mistakes, next up; linking in makefile

+++ NNV/nnv.c
@@ -4,16 +4,16 @@
 static float temp_nnv = 0;
 
 
-float nnv_finite(float* k_arr, float g, uint8_t n_e, float r){
+float nnv_finite(float* k_arr, float g, int n_e, float r){
     float nnv = -g;
-    for(uint8_t i = 0; i < n_e; i++){
+    for(int i = 0; i < n_e; i++){
         nnv += 1 / pow(1+r, 1+i) * k_arr[i];
     }
 
     return nnv;
 }
 
-float nnv_ir_finite(float* k_arr, float g, uint8_t n_e){
+float nnv_ir_finite(float* k_arr, float g, int n_e){
     float temp_r = 0.5;
     temp_nnv = nnv_finite(k_arr, g, n_e, temp_r);
